Allow array of certs from YAML config
The `https.createServer` API allows cert to be an array, to allow for multiple certificates (including intermediate certs in a chain). Node's https server doesn't appear to support the more common pattern of multiple `BEGIN CERTIFICATE`/`END CERTIFICATE` blocks in the same file. In order to specify the sufficient chain to be trusted, some users may need to provide the cert parameter as an array.